Fixing bug in wu.zipWith
Reducing the sequence of items isn't the actual behaviour of zipWith
in Haskell. This thing is more easily observable when checking the
type of zipWith3:

	zipWith3 :: (a -> b -> c -> d) -> [a] -> [b] -> [c] -> [d]

The callback function takes three arguments.

There remains one problem in the current implementation. There's no
way to pass a context object in which to evaluate the callback function.
Maybe a zipWithContext should be created?